Add:   - DurationParser containing functions to transform string "11:00" to Duration and back Changes:   - Lectures to make use of duration and affected classes/tests

+++ db/CreateProjectTable.sql
@@ -74,8 +74,8 @@ CREATE TABLE IF NOT EXISTS LECTURE (
     room_id INT NOT NULL REFERENCES ROOM(id) ON DELETE CASCADE,
     class_type VARCHAR(20) CHECK (class_type IN ('theoretical', 'practical', 'theoretical_practical')),
     week_day int CHECK(week_day > 0 and week_day < 8),
-    start_time bigint NOT NULL,
-    end_time bigint NOT NULL,
+    start_time int NOT NULL,
+    end_time int NOT NULL,
     CHECK (end_time > start_time)
 );
